+++ index.js
@@ -1,18 +1,11 @@
-import { getComments, addComment } from "./api.js";
-import {
-  //deleteLastComment, deleteComment 
-} from "./formComments.js";
+import { getComments, addComment,
+  // deleteComment 
+  } from "./api.js";
 import { myDate, secureInput } from "./optionalFunction.js";
 import { renderLoginComponent } from "./components/login-component.js"
 
-// export const buttonElement = document.querySelector('button.add-form-button');
 export const listElement = document.querySelector('.comments');
-// export const inputNameElement = document.querySelector('.add-form-name');
-// export const textareaElement = document.querySelector('.add-form-text'); 
-//export const loaderCommentsElement = document.getElementById('loaderComments');
-// export const addFormElement = document.querySelector('.add-form')
 export const loaderCommentsElement = document.getElementById('loaderComments');
-//массив comments
 
 export let comments = [];
 
@@ -29,7 +22,7 @@ export const getFetchPromise = () => {
       const appComments = responseData.comments.map((comment) => {
         return {
           name: comment.author.name,
-          date: myDate(new Date()),
+          date: new Date(comment.date),
           text: comment.text,
           likes: comment.likes,
           isLike: false,
@@ -37,8 +30,8 @@ export const getFetchPromise = () => {
         };
       })
       comments = appComments;
-     
-     
+
+
       loaderCommentsElement.classList.add('-display-none');
       renderComments(listElement);
     }).catch((error) => {
@@ -46,20 +39,6 @@ export const getFetchPromise = () => {
       console.warn(error);
     });
 };
-
-// export const postFetchPromise = () => {
-//   const inputNameElement = document.querySelector('.add-form-name');
-//   const textareaElement = document.querySelector('.add-form-text');
-//   addComment({
-//     name: secureInput(inputNameElement.value),
-//     date: myDate(new Date),
-//     text: secureInput(textareaElement.value),
-//     forceError: false,
-//     token,
-//   }).then(() => {
-//     return getFetchPromise();
-//   })
-// }
 
 export const renderComments = () => {
   const commentHtml = comments.map((comment, index) => {
@@ -81,7 +60,7 @@ export const renderComments = () => {
 
     ${comment.isEdit ? `<button class="save-button button-comment" data-index='${index}'>Сохранить</button>` : `<button class="edit-button button-comment" data-index='${index}'>Редактировать</button>`}
     
-    <button class = "delete-button button-comment" data-index='${index}'>Удалить комментарий</button>
+   <button class = "delete-button button-comment" data-id='${comment.id}'>Удалить комментарий</button>
    </div>
   </div>
 </li>`
@@ -121,8 +100,8 @@ export const renderComments = () => {
 
   if (!token) {
     renderLoginComponent({
-      appEl, 
-      commentHtml, 
+      appEl,
+      commentHtml,
       setToken: (newToken) => {
         token = newToken;
       },
@@ -131,44 +110,7 @@ export const renderComments = () => {
     return;
   }
 
-
   appEl.innerHTML = appHtml;
-
-
-  //   const appHtml = `
-  //   <section id="loaderComments" class="loader -display-none">
-  //     <h4 id="loaderText" class="text-loader">Комментарии загружаются...</h4>
-  //   </section>
-  // <div class="container">
-  //   <ul class="comments">
-  //   <!-- список рендерится из js !!!!!!!-->
-  //   ${commentHtml}
-  //   </ul>
-  //   <div class="add-form">
-  //     <input
-  //       type="text"
-  //       class="add-form-name"
-  //       placeholder="Введите ваше имя"
-  //       value=""
-  //     />
-  //     <textarea
-  //       type="textarea"
-  //       class="add-form-text"
-  //       placeholder="Введите ваш комментарий"
-  //       rows="4"
-  //       value = ""
-  //     ></textarea>
-  //     <div class="add-form-row">
-  //       <button class="add-form-button">Написать</button>
-  //       <button class="add-form-button add-form-button--remove">Удалить последний комментарий</button>
-  //     </div>
-  //   </div>
-  // </div>`;
-
-
-
-  // appEl.innerHTML = appHtml;
-
 
   const buttonElement = document.querySelector('button.add-form-button');
   const listElement = document.querySelector('.comments');
@@ -188,17 +130,13 @@ export const renderComments = () => {
 
     buttonElement.disabled = true;
     buttonElement.textContent = "Добавляется..."
-    addFormElement.classList.add('-display-block')
-
-
-
-    //postFetchPromise()
+    addFormElement.classList.add('-display-block');
 
     addComment({
       name: secureInput(inputNameElement.value),
-      date: 
-      //new Date(),
-      myDate(new Date),
+      date:
+        //new Date(),
+        myDate(new Date()),
       text: secureInput(textareaElement.value),
       forceError: false,
       token,
@@ -289,6 +227,57 @@ export const renderComments = () => {
     }
   };
 
+  const answerQuoteToComment = () => {
+    const commentListItems = document.querySelectorAll('.comment');
+    for (const commentListItem of commentListItems) {
+      commentListItem.addEventListener('click', () => {
+        const userName = commentListItem.dataset.name;
+        const userComment = commentListItem.dataset.comment;
+        textareaElement.value = `*_${userName}: \n${userComment}__*`;
+      })
+    }
+  }
+
+  const deleteComment = () => {
+    const deleteButtonElements = document.querySelectorAll('.delete-button');
+    for (const deleteButtonElement of deleteButtonElements) {
+      deleteButtonElement.addEventListener('click', (e) => {
+        e.stopPropagation();
+        const index = deleteButtonElement.dataset.index;
+        comments.splice(index, 1);
+        renderComments(listElement);
+      });
+    }
+  };
+
+  //Удаление комментария из API(не получается)
+
+  // const deleteButtonElements = document.querySelectorAll('.delete-button');
+  // for (const deleteButtonElement of deleteButtonElements) {
+  //   deleteButtonElement.addEventListener('click', (e) => {
+  //     e.stopPropagation();
+  //     const id = deleteButtonElement.dataset.id;
+
+  //     deleteComment({ token, id })
+  //     .then((responseData) => {
+  //       comments = responseData.comments;
+  //       //renderComments(listElement);
+  //       //getComments();
+  //       renderComments(listElement);
+  //     });
+  //   });
+  // }
+
+  const deleteLastComment = () => {
+    const deleteLastButtonElement = document.querySelector('.add-form-button--remove');
+
+    deleteLastButtonElement.addEventListener('click', () => {
+      comments.pop();
+      renderComments(listElement);
+    });
+  };
+
+
   const validateButton = () => {
     if (!inputNameElement.value || !textareaElement.value) {
       buttonElement.disabled = true;
@@ -313,45 +302,18 @@ export const renderComments = () => {
       }
     });
   }
-
   pushEnter();
-  //pushNewComment();
   initChangeLikeButtonListeners();
   initEditButtonListeners();
   initDisabledButtonElement();
-  //deleteComment();
-  //answerToComment(); ДЗ 2.11
-  //answerQuoteToComment(); // ДЗ со звездочкой 2.11
+  answerQuoteToComment();
+  deleteComment();
+  deleteLastComment();
 }
 
-//getFetchPromise();
 renderComments(listElement)
-
-
-//Функция POST
-
-// export const postFetchPromise = () => {
-//   addComment({
-//     name: secureInput(inputNameElement.value),
-//     date: myDate(new Date),
-//     text: secureInput(textareaElement.value),
-//     likes: 0,
-//     isLike: false,
-//     forceError: false,
-//     token
-//   }).then(() => {
-//     return getFetchPromise();
-//   })
-// }
-
-//validateButton()
-//initDisabledButtonElement()
-//deleteLastComment();
-//deleteComment();
 getFetchPromise();
-//pushEnter();
-//pushNewComment();
-renderComments(listElement);
+
