+++ optionalFunction.js
@@ -1,7 +1,7 @@
 //рендер даты
 
-export const myDate = () => {
-    const getDate = new Date();
+export const myDate = (date) => {
+    const getDate = new Date(date);
     const options = {
       year: "2-digit",
       month: "2-digit",
@@ -14,6 +14,20 @@ export const myDate = () => {
     options.hour = "2-digit";
     return getDate.toLocaleDateString("ru-Ru", options).split(', ').join(' ');
   }
+// export const myDate = () => {
+//     const getDate = new Date();
+//     const options = {
+//       year: "2-digit",
+//       month: "2-digit",
+//       day: "2-digit",
+//       timezone: "UTC",
+//       hour: "numeric",
+//       minute: "2-digit"
+//     };
+    
+//     options.hour = "2-digit";
+//     return getDate.toLocaleDateString("ru-Ru", options).split(', ').join(' ');
+//   }
 
   //обезопасить ввод данных пользователя
   
@@ -23,7 +37,7 @@ export const myDate = () => {
   //   //.replaceAll("&", "&amp;")
   //   //.replaceAll('"', "&quot;");
   // }
-  export function secureInput(safeText) {
+  export function secureInput(safeText){
     return safeText.replaceAll("<", "&lt;")
       .replaceAll(">", "&gt;")
     //.replaceAll("&", "&amp;")
